Added a base class for QueuedStatusPush
- We can re-use this for HTTPStatus and probably also AutoBahn
